refactoring: controllers
return AuthenticationController, PortalController and SecurityController
refactoing settingscontollers
