+++ web/ASC.Web.Api/Api/Settings/SettingsController.cs
@@ -3,71 +3,98 @@
 namespace ASC.Web.Api.Controllers.Settings;
 public class SettingsController: BaseSettingsController
 {
-    public SettingsController(IOptionsMonitor<ILog> option,
-    MessageService messageService,
-    StudioNotifyService studioNotifyService,
-    ApiContext apiContext,
-    UserManager userManager,
-    TenantManager tenantManager,
-    TenantExtra tenantExtra,
-    TenantStatisticsProvider tenantStatisticsProvider,
-    AuthContext authContext,
-    CookiesManager cookiesManager,
-    WebItemSecurity webItemSecurity,
-    StudioNotifyHelper studioNotifyHelper,
-    LicenseReader licenseReader,
-    PermissionContext permissionContext,
-    SettingsManager settingsManager,
-    TfaManager tfaManager,
-    WebItemManager webItemManager,
-    WebItemManagerSecurity webItemManagerSecurity,
-    TenantInfoSettingsHelper tenantInfoSettingsHelper,
-    TenantWhiteLabelSettingsHelper tenantWhiteLabelSettingsHelper,
-    StorageHelper storageHelper,
-    TenantLogoManager tenantLogoManager,
-    TenantUtil tenantUtil,
-    CoreBaseSettings coreBaseSettings,
-    CommonLinkUtility commonLinkUtility,
-    ColorThemesSettingsHelper colorThemesSettingsHelper,
-    IConfiguration configuration,
-    SetupInfo setupInfo,
-    BuildVersion buildVersion,
-    DisplayUserSettingsHelper displayUserSettingsHelper,
-    StatisticManager statisticManager,
-    IPRestrictionsService iPRestrictionsService,
-    CoreConfiguration coreConfiguration,
-    MessageTarget messageTarget,
-    StudioSmsNotificationSettingsHelper studioSmsNotificationSettingsHelper,
-    CoreSettings coreSettings,
-    StorageSettingsHelper storageSettingsHelper,
-    IWebHostEnvironment webHostEnvironment,
-    IServiceProvider serviceProvider,
-    EmployeeWraperHelper employeeWraperHelper,
-    ConsumerFactory consumerFactory,
-    SmsProviderManager smsProviderManager,
-    TimeZoneConverter timeZoneConverter,
-    CustomNamingPeople customNamingPeople,
-    IPSecurity.IPSecurity ipSecurity,
-    IMemoryCache memoryCache,
-    ProviderManager providerManager,
-    FirstTimeTenantSettings firstTimeTenantSettings,
-    ServiceClient serviceClient,
-    TelegramHelper telegramHelper,
-    StorageFactory storageFactory,
-    UrlShortener urlShortener,
-    EncryptionServiceClient encryptionServiceClient,
-    EncryptionSettingsHelper encryptionSettingsHelper,
-    BackupAjaxHandler backupAjaxHandler,
-    ICacheNotify<DeleteSchedule> cacheDeleteSchedule,
-    EncryptionWorker encryptionWorker,
-    PasswordHasher passwordHasher,
-    PaymentManager paymentManager,
-    Constants constants,
-    InstanceCrypto instanceCrypto,
-    Signature signature,
-    DbWorker dbWorker,
-    IHttpClientFactory clientFactory) : base(option, messageService, studioNotifyService, apiContext, userManager, tenantManager, tenantExtra, tenantStatisticsProvider, authContext, cookiesManager, webItemSecurity, studioNotifyHelper, licenseReader, permissionContext, settingsManager, tfaManager, webItemManager, webItemManagerSecurity, tenantInfoSettingsHelper, tenantWhiteLabelSettingsHelper, storageHelper, tenantLogoManager, tenantUtil, coreBaseSettings, commonLinkUtility, colorThemesSettingsHelper, configuration, setupInfo, buildVersion, displayUserSettingsHelper, statisticManager, iPRestrictionsService, coreConfiguration, messageTarget, studioSmsNotificationSettingsHelper, coreSettings, storageSettingsHelper, webHostEnvironment, serviceProvider, employeeWraperHelper, consumerFactory, smsProviderManager, timeZoneConverter, customNamingPeople, ipSecurity, memoryCache, providerManager, firstTimeTenantSettings, serviceClient, telegramHelper, storageFactory, urlShortener, encryptionServiceClient, encryptionSettingsHelper, backupAjaxHandler, cacheDeleteSchedule, encryptionWorker, passwordHasher, paymentManager, constants, instanceCrypto, signature, dbWorker, clientFactory)
+    private Tenant Tenant { get { return _apiContext.Tenant; } }
+
+    private readonly MessageService _messageService;
+    private readonly ConsumerFactory _consumerFactory;
+    private readonly TimeZoneConverter _timeZoneConverter;
+    private readonly CustomNamingPeople _customNamingPeople;
+    private readonly ProviderManager _providerManager;
+    private readonly FirstTimeTenantSettings _firstTimeTenantSettings;
+    private readonly UserManager _userManager;
+    private readonly TenantManager _tenantManager;
+    private readonly TenantExtra _tenantExtra;
+    private readonly TenantStatisticsProvider _tenantStatisticsProvider;
+    private readonly AuthContext _authContext;
+    private readonly PermissionContext _permissionContext;
+    private readonly SettingsManager _settingsManager;
+    private readonly WebItemManagerSecurity _webItemManagerSecurity;
+    private readonly TenantInfoSettingsHelper _tenantInfoSettingsHelper;
+    private readonly TenantUtil _tenantUtil;
+    private readonly CoreBaseSettings _coreBaseSettings;
+    private readonly CommonLinkUtility _commonLinkUtility;
+    private readonly ColorThemesSettingsHelper _colorThemesSettingsHelper;
+    private readonly IConfiguration _configuration;
+    private readonly SetupInfo _setupInfo;
+    private readonly StatisticManager _statisticManager;
+    private readonly CoreConfiguration _coreConfiguration;
+    private readonly UrlShortener _urlShortener;
+    private readonly PasswordHasher _passwordHasher;
+    private readonly ILog _log;
+    private readonly TelegramHelper _telegramHelper;
+    private readonly Constants _constants;
+
+    public SettingsController(
+        IOptionsMonitor<ILog> option,
+        MessageService messageService,
+        ApiContext apiContext,
+        UserManager userManager,
+        TenantManager tenantManager,
+        TenantExtra tenantExtra,
+        TenantStatisticsProvider tenantStatisticsProvider,
+        AuthContext authContext,
+        PermissionContext permissionContext,
+        SettingsManager settingsManager,
+        WebItemManager webItemManager,
+        WebItemManagerSecurity webItemManagerSecurity,
+        TenantInfoSettingsHelper tenantInfoSettingsHelper,
+        TenantUtil tenantUtil,
+        CoreBaseSettings coreBaseSettings,
+        CommonLinkUtility commonLinkUtility,
+        ColorThemesSettingsHelper colorThemesSettingsHelper,
+        IConfiguration configuration,
+        SetupInfo setupInfo,
+        StatisticManager statisticManager,
+        CoreConfiguration coreConfiguration,
+        ConsumerFactory consumerFactory,
+        TimeZoneConverter timeZoneConverter,
+        CustomNamingPeople customNamingPeople,
+        IMemoryCache memoryCache,
+        ProviderManager providerManager,
+        FirstTimeTenantSettings firstTimeTenantSettings,
+        TelegramHelper telegramHelper,
+        UrlShortener urlShortener,
+        PasswordHasher passwordHasher,
+        Constants constants) : base(apiContext, memoryCache, webItemManager)
     {
+        _log = option.Get("ASC.Api");
+        _consumerFactory = consumerFactory;
+        _timeZoneConverter = timeZoneConverter;
+        _customNamingPeople = customNamingPeople;
+        _providerManager = providerManager;
+        _firstTimeTenantSettings = firstTimeTenantSettings;
+        _messageService = messageService;
+        _userManager = userManager;
+        _tenantManager = tenantManager;
+        _tenantExtra = tenantExtra;
+        _tenantStatisticsProvider = tenantStatisticsProvider;
+        _authContext = authContext;
+        _permissionContext = permissionContext;
+        _settingsManager = settingsManager;
+        _webItemManagerSecurity = webItemManagerSecurity;
+        _tenantInfoSettingsHelper = tenantInfoSettingsHelper;
+        _tenantUtil = tenantUtil;
+        _coreBaseSettings = coreBaseSettings;
+        _commonLinkUtility = commonLinkUtility;
+        _colorThemesSettingsHelper = colorThemesSettingsHelper;
+        _configuration = configuration;
+        _setupInfo = setupInfo;
+        _statisticManager = statisticManager;
+        _coreConfiguration = coreConfiguration;
+        _passwordHasher = passwordHasher;
+        _urlShortener = urlShortener;
+        _telegramHelper = telegramHelper;
+        _constants = constants;
     }
 
     [Read("", Check = false)]
