+++ web/ASC.Web.Api/Api/Settings/StorageController.cs
@@ -4,71 +4,66 @@ namespace ASC.Web.Api.Controllers.Settings;
 
 public class StorageController: BaseSettingsController
 {
-    public StorageController(IOptionsMonitor<ILog> option,
-    MessageService messageService,
-    StudioNotifyService studioNotifyService,
-    ApiContext apiContext,
-    UserManager userManager,
-    TenantManager tenantManager,
-    TenantExtra tenantExtra,
-    TenantStatisticsProvider tenantStatisticsProvider,
-    AuthContext authContext,
-    CookiesManager cookiesManager,
-    WebItemSecurity webItemSecurity,
-    StudioNotifyHelper studioNotifyHelper,
-    LicenseReader licenseReader,
-    PermissionContext permissionContext,
-    SettingsManager settingsManager,
-    TfaManager tfaManager,
-    WebItemManager webItemManager,
-    WebItemManagerSecurity webItemManagerSecurity,
-    TenantInfoSettingsHelper tenantInfoSettingsHelper,
-    TenantWhiteLabelSettingsHelper tenantWhiteLabelSettingsHelper,
-    StorageHelper storageHelper,
-    TenantLogoManager tenantLogoManager,
-    TenantUtil tenantUtil,
-    CoreBaseSettings coreBaseSettings,
-    CommonLinkUtility commonLinkUtility,
-    ColorThemesSettingsHelper colorThemesSettingsHelper,
-    IConfiguration configuration,
-    SetupInfo setupInfo,
-    BuildVersion buildVersion,
-    DisplayUserSettingsHelper displayUserSettingsHelper,
-    StatisticManager statisticManager,
-    IPRestrictionsService iPRestrictionsService,
-    CoreConfiguration coreConfiguration,
-    MessageTarget messageTarget,
-    StudioSmsNotificationSettingsHelper studioSmsNotificationSettingsHelper,
-    CoreSettings coreSettings,
-    StorageSettingsHelper storageSettingsHelper,
-    IWebHostEnvironment webHostEnvironment,
-    IServiceProvider serviceProvider,
-    EmployeeWraperHelper employeeWraperHelper,
-    ConsumerFactory consumerFactory,
-    SmsProviderManager smsProviderManager,
-    TimeZoneConverter timeZoneConverter,
-    CustomNamingPeople customNamingPeople,
-    IPSecurity.IPSecurity ipSecurity,
-    IMemoryCache memoryCache,
-    ProviderManager providerManager,
-    FirstTimeTenantSettings firstTimeTenantSettings,
-    ServiceClient serviceClient,
-    TelegramHelper telegramHelper,
-    StorageFactory storageFactory,
-    UrlShortener urlShortener,
-    EncryptionServiceClient encryptionServiceClient,
-    EncryptionSettingsHelper encryptionSettingsHelper,
-    BackupAjaxHandler backupAjaxHandler,
-    ICacheNotify<DeleteSchedule> cacheDeleteSchedule,
-    EncryptionWorker encryptionWorker,
-    PasswordHasher passwordHasher,
-    PaymentManager paymentManager,
-    Constants constants,
-    InstanceCrypto instanceCrypto,
-    Signature signature,
-    DbWorker dbWorker,
-    IHttpClientFactory clientFactory) : base(option, messageService, studioNotifyService, apiContext, userManager, tenantManager, tenantExtra, tenantStatisticsProvider, authContext, cookiesManager, webItemSecurity, studioNotifyHelper, licenseReader, permissionContext, settingsManager, tfaManager, webItemManager, webItemManagerSecurity, tenantInfoSettingsHelper, tenantWhiteLabelSettingsHelper, storageHelper, tenantLogoManager, tenantUtil, coreBaseSettings, commonLinkUtility, colorThemesSettingsHelper, configuration, setupInfo, buildVersion, displayUserSettingsHelper, statisticManager, iPRestrictionsService, coreConfiguration, messageTarget, studioSmsNotificationSettingsHelper, coreSettings, storageSettingsHelper, webHostEnvironment, serviceProvider, employeeWraperHelper, consumerFactory, smsProviderManager, timeZoneConverter, customNamingPeople, ipSecurity, memoryCache, providerManager, firstTimeTenantSettings, serviceClient, telegramHelper, storageFactory, urlShortener, encryptionServiceClient, encryptionSettingsHelper, backupAjaxHandler, cacheDeleteSchedule, encryptionWorker, passwordHasher, paymentManager, constants, instanceCrypto, signature, dbWorker, clientFactory)
+    private Tenant Tenant { get { return _apiContext.Tenant; } }
+
+    private readonly MessageService _messageService;
+    private readonly StudioNotifyService _studioNotifyService;
+    private readonly IWebHostEnvironment _webHostEnvironment;
+    private readonly ConsumerFactory _consumerFactory;
+    private readonly TenantManager _tenantManager;
+    private readonly TenantExtra _tenantExtra;
+    private readonly PermissionContext _permissionContext;
+    private readonly SettingsManager _settingsManager;
+    private readonly CoreBaseSettings _coreBaseSettings;
+    private readonly CommonLinkUtility _commonLinkUtility;
+    private readonly StorageSettingsHelper _storageSettingsHelper;
+    private readonly ServiceClient _serviceClient;
+    private readonly EncryptionServiceClient _encryptionServiceClient;
+    private readonly EncryptionSettingsHelper _encryptionSettingsHelper;
+    private readonly BackupAjaxHandler _backupAjaxHandler;
+    private readonly ICacheNotify<DeleteSchedule> _cacheDeleteSchedule;
+    private readonly EncryptionWorker _encryptionWorker;
+    private readonly ILog _log;
+
+    public StorageController(
+        IOptionsMonitor<ILog> option,
+        MessageService messageService,
+        StudioNotifyService studioNotifyService,
+        ApiContext apiContext,
+        TenantManager tenantManager,
+        TenantExtra tenantExtra,
+        PermissionContext permissionContext,
+        SettingsManager settingsManager,
+        WebItemManager webItemManager,
+        CoreBaseSettings coreBaseSettings,
+        CommonLinkUtility commonLinkUtility,
+        StorageSettingsHelper storageSettingsHelper,
+        IWebHostEnvironment webHostEnvironment,
+        ConsumerFactory consumerFactory,
+        IMemoryCache memoryCache,
+        EncryptionServiceClient encryptionServiceClient,
+        EncryptionSettingsHelper encryptionSettingsHelper,
+        BackupAjaxHandler backupAjaxHandler,
+        ICacheNotify<DeleteSchedule> cacheDeleteSchedule,
+        EncryptionWorker encryptionWorker) : base(apiContext, memoryCache, webItemManager)
     {
+        _log = option.Get("ASC.Api");
+        _webHostEnvironment = webHostEnvironment;
+        _consumerFactory = consumerFactory;
+        _messageService = messageService;
+        _studioNotifyService = studioNotifyService;
+        _tenantManager = tenantManager;
+        _tenantExtra = tenantExtra;
+        _permissionContext = permissionContext;
+        _settingsManager = settingsManager;
+        _coreBaseSettings = coreBaseSettings;
+        _commonLinkUtility = commonLinkUtility;
+        _storageSettingsHelper = storageSettingsHelper;
+        _encryptionServiceClient = encryptionServiceClient;
+        _encryptionSettingsHelper = encryptionSettingsHelper;
+        _backupAjaxHandler = backupAjaxHandler;
+        _cacheDeleteSchedule = cacheDeleteSchedule;
+        _encryptionWorker = encryptionWorker;
     }
 
     [Read("storage")]
