+++ web/ASC.Web.Api/GlobalUsings.cs
@@ -59,9 +59,6 @@ global using ASC.Security.Cryptography;
 global using ASC.Web.Api;
 global using ASC.Web.Api.ApiModel;
 global using ASC.Web.Api.Controllers;
-global using ASC.Web.Api.Controllers.Authentication;
-global using ASC.Web.Api.Controllers.Portal;
-global using ASC.Web.Api.Controllers.Security;
 global using ASC.Web.Api.Controllers.Settings;
 global using ASC.Web.Api.Core;
 global using ASC.Web.Api.Models;
